Better Bookshelves - User bookshelves now use SSR for display - Links in the profile bar now use safe routing - Cleaned up the test projects somewhat

+++ Ogma3/Pages/Shelves/Bookshelf.cshtml.cs
@@ -14,7 +14,7 @@ public sealed class Bookshelf(ApplicationDbContext context) : PageModel
 
 	public sealed record BookshelfDetails(string Name, string Description, string? Color, string? IconName, IEnumerable<StoryCard> Stories);
 
-	public async Task<IActionResult> OnGetAsync(int id, string? slug)
+	public async Task<IActionResult> OnGetAsync(long id, string? slug)
 	{
 		var uid = User.GetNumericId();
